Moves around files into an OCD format
