+++ src/dm_preprocessor/directive/ifndef.rs
@@ -1,5 +1,5 @@
 use crate::{
-    dm_preprocessor::{token_handling::DmToken, DmPreProcessor},
+    dm_preprocessor::{token::dm_token::DmToken, DmPreProcessor},
     util::ParseError,
 };
