+++ src/dm_preprocessor/directive/include.rs
@@ -1,7 +1,7 @@
 use log::{error, trace};
 
 use crate::{
-    dm_preprocessor::{token_handling::DmToken, DmPreProcessor},
+    dm_preprocessor::{token::dm_token::DmToken, DmPreProcessor},
     util::ParseError,
 };
