+++ src/dm_preprocessor/directive/mod.rs
@@ -2,7 +2,7 @@ use log::{error, warn};
 
 use crate::util::ParseError;
 
-use super::{token_handling::DmToken, DmPreProcessor};
+use super::{token::dm_token::DmToken, DmPreProcessor};
 
 pub mod define;
 pub mod error;
