--- src/dm_preprocessor/tests/mod.rs
@@ -1,1 +1,0 @@
-mod token_handling;
