+++ src/dm_preprocessor/token/count_backslashes.rs
@@ -1,0 +1,11 @@
+pub fn count_backslashes(string: &str) -> usize {
+    let mut count = 0;
+    for char in string.chars().rev() {
+        if char == '\\' {
+            count += 1;
+        } else {
+            break;
+        }
+    }
+    count
+}
