+++ src/dm_preprocessor/token/tokenize_state.rs
@@ -1,54 +1,6 @@
-use std::fmt::Display;
+use log::trace;
 
-use log::{error, trace};
-
-use crate::util::{
-    condense_brackets::condense_braces,
-    condense_lines::condense_lines,
-    count_backslashes,
-    determine_token_action::{determine_token_action, TokenAction},
-};
-
-use super::DmPreProcessor;
-
-#[derive(Debug, Clone)]
-pub struct DmToken {
-    value: String,
-}
-
-impl Display for DmToken {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.value)
-    }
-}
-
-impl DmToken {
-    pub fn new(value: String) -> Self {
-        Self { value }
-    }
-
-    pub fn value(&self) -> &str {
-        &self.value
-    }
-}
-
-impl From<&str> for DmToken {
-    fn from(value: &str) -> Self {
-        Self::new(value.into())
-    }
-}
-
-impl From<String> for DmToken {
-    fn from(value: String) -> Self {
-        Self::from(value.as_str())
-    }
-}
-
-impl PartialEq for DmToken {
-    fn eq(&self, other: &Self) -> bool {
-        self.value == other.value
-    }
-}
+use super::{count_backslashes::count_backslashes, dm_token::DmToken};
 
 #[derive(Debug, Default)]
 pub struct TokenizeState {
@@ -206,81 +158,5 @@ impl TokenizeState {
             return false;
         }
         count_backslashes(last.unwrap().value()) % 2 == 1
-    }
-}
-
-impl DmPreProcessor {
-    pub fn tokenize(&mut self, lines: &[impl Into<String> + Clone]) -> Vec<DmToken> {
-        let condensed_lines: Vec<String> = condense_lines(lines);
-        let condensed_lines = condense_braces(&condensed_lines);
-        let mut tokens: Vec<DmToken> = vec![];
-
-        for line in condensed_lines {
-            let mut token = String::new();
-            self.tokenize_state.set_in_preprocessor(false);
-            self.tokenize_state.set_comment_single(false);
-
-            trace!("Tokenizing line: `{}`", line);
-            for char in line.chars() {
-                trace!("Char: `{}`", char.escape_debug());
-                let next_action = determine_token_action(&mut self.tokenize_state, char, &token);
-                match next_action {
-                    TokenAction::StartNewToken => {
-                        if !token.is_empty() {
-                            self.tokenize_state.add_line_token(token);
-                        }
-                        token = char.to_string();
-                    }
-                    TokenAction::ContinueToken => {
-                        token.push(char);
-                    }
-                    TokenAction::EndToken => {
-                        token.push(char);
-                        self.tokenize_state.add_line_token(token);
-                        token = String::new();
-                    }
-                    TokenAction::IsolateToken => {
-                        if !token.is_empty() {
-                            self.tokenize_state.add_line_token(token);
-                        }
-                        self.tokenize_state.add_line_token(char.to_string());
-                        token = String::new();
-                    }
-                    _ => {
-                        error!(
-                            "Unexpected token action `{}` with char {}",
-                            next_action, char
-                        );
-                        panic!();
-                    }
-                }
-            }
-
-            if !token.is_empty() {
-                self.tokenize_state.add_line_token(token);
-            }
-            self.tokenize_state.add_line_token("\n");
-            tokens.append(&mut self.tokenize_state.finalize_line_tokens());
-            if self.tokenize_state.in_quote().is_some() && !self.tokenize_state.in_preprocessor() {
-                error!(
-                    "Unterminated quote `{}` in line `{}`",
-                    self.tokenize_state.in_quote().unwrap(),
-                    line
-                );
-                panic!();
-            }
-
-            if self.tokenize_state.unmatched_brackets() {
-                error!("Unmatched brackets in line `{}`", line);
-                panic!();
-            }
-
-            if self.tokenize_state.in_string_interop() {
-                error!("Unmatched string interop in line `{}`", line);
-                panic!();
-            }
-        }
-
-        tokens
     }
 }
